Word detection demo
Added code that scans word-detection folder and loads all .txt files from there. Words in these files are scanned in everymessage. If word is found in a message this is then logged. (Logging messages not ready yet this version only has detection for them)

Added a new command `!m setswearchannel` When this command is used it marks the channel it was used in as a log channel for messages found containing bad words.